Fix linux/386 mage package (#9622)
Converts `stats.Ctim.sec` and `stats.Ctim.nsec` to int64 explicitly to accommodate 32-bit systems.

+++ x-pack/auditbeat/module/system/user/user.go
@@ -522,7 +522,7 @@ func (ms *MetricSet) haveFilesChanged() (bool, error) {
 			return true, errors.Wrapf(err, "failed to stat %v", path)
 		}
 
-		ctime := time.Unix(stats.Ctim.Sec, stats.Ctim.Nsec)
+		ctime := time.Unix(int64(stats.Ctim.Sec), int64(stats.Ctim.Nsec))
 		if ms.lastRead.Before(ctime) {
 			ms.log.Debugf("File changed: %v (lastRead=%v, ctime=%v)", path, ms.lastRead, ctime)
